zjl@2017/3/30 R03.128

+++ SrcBxxh/l0bsp_stm32/bsp_spi.c
@@ -526,7 +526,7 @@ uint8_t ihu_bsp_stm32_spi_nrf24l01_read_write_byte(uint8_t byte)
   uint8_t d_read = 0, d_send = byte;
   if(HAL_SPI_TransmitReceive(&IHU_BSP_STM32_SPI_RFID_NRF24L01_HANDLER, &d_send,&d_read,1,IHU_BSP_STM32_SPI_TX_RX_MAX_DELAY)!=HAL_OK)
   {
-    d_read = IHU_BSP_STM32_SPI_TX_RX_MAX_DELAY;
+    d_read = (uint8_t)IHU_BSP_STM32_SPI_TX_RX_MAX_DELAY;
   }
   return d_read; 
 }
